requirements.txt
opencv-python
mediapipe
numpy
Pillow
tensorflow
